[FIX] Plugin: break cyclic dependance of visibility plugins
The HeaderVisibility and HeaderGrouping plugins depend on each other's
getters, therefore breaking the dependency chain.

Task: 4908195
