+++ src/functions/module_math.ts
@@ -1416,7 +1416,7 @@ export const SUBTOTAL = {
         if (this.getters.isRowFiltered(sheetId, row)) {
           continue;
         }
-        if (!acceptHiddenCells && this.getters.isRowHiddenByUser(sheetId, row)) {
+        if (!acceptHiddenCells && this.getters.isRowHiddenOrFoldedByUser(sheetId, row)) {
           continue;
         }
