+++ src/types/core_getters.ts
@@ -9,6 +9,7 @@ import { FigurePlugin } from "../plugins/core/figures";
 import { HeaderGroupingPlugin } from "../plugins/core/header_grouping";
 import { HeaderSizePlugin } from "../plugins/core/header_size";
 import { HeaderVisibilityPlugin } from "../plugins/core/header_visibility";
+import { HiddenHeaderPlugin } from "../plugins/core/hidden_headers";
 import { ImagePlugin } from "../plugins/core/image";
 import { MergePlugin } from "../plugins/core/merge";
 import { NamedRangesPlugin } from "../plugins/core/named_range";
@@ -71,6 +72,7 @@ export type PluginGetters<
 type RangeAdapterGetters = Pick<RangeAdapterPlugin, GetterNames<typeof RangeAdapterPlugin>>;
 export type CoreGetters = PluginGetters<typeof SheetPlugin> &
   PluginGetters<typeof HeaderSizePlugin> &
+  PluginGetters<typeof HiddenHeaderPlugin> &
   PluginGetters<typeof HeaderVisibilityPlugin> &
   PluginGetters<typeof CellPlugin> &
   PluginGetters<typeof DefaultPlugin> &
